Add debug msg to assertion in python node.handle
There are a number of students reporting that they're seeing errors in
which the implicit `recv_multipart` that ZMQ calls for `on_recv`
callbacks is returning a message with too many parts. One student
suggested that his messages may be getting appended to one another when
sent in quick succession. Adding a debug message is not necessarily a
safe call for tracking this down -- the timing changes due to debug mode
seem to make the errors disappear for everyone. Instead of shuffling the
debug, add a quick format string to the assertion itself, so it will
always give us the full information.